requirements.txt
streamlit>=1.30.0
htbuilder>=0.9.0
langchain>=0.2.0
langchain-core>=0.2.0
langchain-community>=0.2.0
langchain-ollama>=0.1.0
langchain-chroma>=0.1.0
langchain-text-splitters>=0.2.0
chromadb>=0.4.22
beautifulsoup4>=4.12.0
requests>=2.31.0
pymupdf>=1.23.0
python-docx>=1.1.0
Pillow>=10.0.0
pytesseract>=0.3.10
lxml>=5.0.0
html2text>=2024.2.26
langdetect>=1.0.9
tiktoken>=0.6.0
